database interaction possibly done
